+++ pyanaconda/task/task.py
@@ -28,7 +28,7 @@ from pyanaconda.isignal import Signal
 from pyanaconda.constants import THREAD_DBUS_TASK
 from pyanaconda.task.task_interface import TaskAlreadyRunningException
 from pyanaconda.threading import threadMgr, AnacondaThread
-from pyanaconda.async_utils import async_action_nowait
+from pyanaconda.core.async_utils import async_action_nowait
 
 __all__ = ['Task', 'TaskAlreadyRunningException']
