+++ pyanaconda/ui/gui/spokes/advstorage/fcoe.py
@@ -22,7 +22,7 @@ from blivet.fcoe import fcoe
 from pyanaconda import constants
 from pyanaconda.threading import threadMgr, AnacondaThread
 from pyanaconda.ui.gui import GUIObject
-from pyanaconda.async_utils import async_action_wait
+from pyanaconda.core.async_utils import async_action_wait
 from pyanaconda.storage_utils import try_populate_devicetree
 from pyanaconda import nm
