+++ pyanaconda/ui/gui/spokes/lib/entropy_dialog.py
@@ -29,7 +29,7 @@ from pyanaconda.i18n import P_
 from pyanaconda.constants import MAX_ENTROPY_WAIT
 from pyanaconda.core.timer import Timer
 from pyanaconda.ui.gui import GUIObject
-from pyanaconda.async_utils import async_action_wait
+from pyanaconda.core.async_utils import async_action_wait
 from blivet.util import get_current_entropy
 
 __all__ = ["run_entropy_dialog"]
