+++ pyanaconda/ui/gui/utils.py
@@ -34,7 +34,7 @@ from contextlib import contextmanager
 
 from pyanaconda.core import glib
 from pyanaconda.threading import threadMgr, AnacondaThread
-from pyanaconda.async_utils import async_action_wait, run_in_loop
+from pyanaconda.core.async_utils import async_action_wait, run_in_loop
 from pyanaconda.constants import NOTICEABLE_FREEZE, PASSWORD_HIDE, PASSWORD_SHOW, \
                                  PASSWORD_HIDE_ICON, PASSWORD_SHOW_ICON
