+++ pyanaconda/ui/gui/xkl_wrapper.py
@@ -41,7 +41,7 @@ from pyanaconda import flags
 from pyanaconda import iutil
 from pyanaconda.constants import DEFAULT_KEYBOARD
 from pyanaconda.keyboard import join_layout_variant, parse_layout_variant, KeyboardConfigError, InvalidLayoutVariantSpec
-from pyanaconda.async_utils import async_action_wait
+from pyanaconda.core.async_utils import async_action_wait
 
 from pyanaconda.anaconda_loggers import get_module_logger
 log = get_module_logger(__name__)
